adding package extension support
