+++ src/plotting.jl
@@ -1,133 +1,11 @@
-const meshpath = artifact"DKTMeshes"
-mni_cortex() = joinpath(meshpath, "meshes/cortex/connectome-cortex.obj")
-fs_cortex() =  joinpath(meshpath, "meshes/cortex/fs-cortex.obj")
-rh_cortex() =  joinpath(meshpath, "meshes/cortex/rh-cortex.obj")
-lh_cortex() =  joinpath(meshpath, "meshes/cortex/lh-cortex.obj")
-
-function set_fig(;resolution::Tuple{Int64, Int64}=(1600,900), view=:front)
-    f = Figure(resolution = resolution)
-    ax = Axis3(f[1,1], aspect = :data, azimuth = View[view]pi, elevation=0.0pi)
-    hidedecorations!(ax)
-    hidespines!(ax)
-    f
-end
-
-RegionDict = Dict(zip([:left, :right, :all, :connectome], [lh_cortex(), rh_cortex(), fs_cortex(), mni_cortex()]))
-
-View = Dict(zip([:right, :front, :left, :back], [0.0, 0.5, 1.0, 1.5]))
-
-function plot_cortex!(region::Symbol=:all; color=(:grey,0.05), transparency::Bool=true, kwargs...)
-    mesh!(load(RegionDict[region]), color=color, transparency=transparency, kwargs...)
-end
-
-function plot_cortex(region::Symbol=:all; resolution=(800, 600), view=:left, color=(:grey,1.0), transparency::Bool=false, kwargs...)
-    f = set_fig(resolution=resolution, view=view)
-    plot_cortex!(region; color, transparency, kwargs...)
-    f
-end
-
-function plot_parc!(parc::Parcellation)
-    colors = distinguishable_colors(length(parc))
-    for (i, roi) in enumerate(parc)
-        plot_roi!(get_id(roi), colors[i])
-    end
-end
-
-function plot_parc(parc::Parcellation; view=:left)
-    f = set_fig(;view=view)
-    plot_parc!(parc)
-    f
-end
-
-function plot_roi!(roi::Int, color=(:grey,1.0); transparency=false)
-    roi = joinpath(meshpath, "meshes/DKT/roi_$(roi).obj")
-    mesh!(load(roi), color=color, transparency=transparency)
-end
-
-function plot_roi!(rois::Vector{Int64}, colors::Vector{Float64}, cmap::ColorScheme; transparency=false)
-    [plot_roi!(roi, get(cmap, color); transparency=transparency) for (roi, color) in zip(rois, colors)]
-end
-
-function plot_roi(rois, colors, cmap; transparency=false, view=:front, resolution=(1600,900))
-    f  = set_fig(;resolution=resolution, view=view)
-    plot_roi!(rois, colors, cmap; transparency=transparency)
-    f
-end
-
-function plot_vertex!(connectome::Connectome, node_size=10, color=(:blue, 0.5), transparency::Bool=true)
-    x, y, z = connectome.parc.x[:], connectome.parc.y[:], connectome.parc.z[:]
-    meshscatter!(x, y, z, markersize=node_size, color=color, transparency=transparency)
-end
-
-function plot_vertices(connectome::Connectome; node_size=1.0, color=(:blue,0.5))
-    f  = set_fig()
-    plot_cortex!(:connectome)
-    plot_vertex!(connectome, node_size, color)
-    f
-end
-
-function plot_connectome(connectome::Connectome; 
-                              edge_weighted=true, 
-                              edge_alpha=false,
-                              edge_map = ColorSchemes.viridis,
-                              min_edge_size = 0.0,
-                              max_edge_size = 10.0,
-                              node_weighted = true,
-                              node_color= (:blue, 0.5),
-                              node_size = 10.0)
-    f = set_fig()
-    plot_cortex!(:connectome)
-    plot_connectome!(connectome; 
-                     edge_weighted=edge_weighted, 
-                     edge_map=edge_map, 
-                     edge_alpha=edge_alpha,
-                     min_edge_size = min_edge_size,
-                     max_edge_size = max_edge_size,
-                     node_weighted=node_weighted,
-                     node_color=node_color,
-                     node_size=node_size)
-    f
-end
-
-
-function plot_connectome!(connectome::Connectome; 
-                              edge_weighted=true,
-                              edge_alpha=false, 
-                              edge_map = ColorSchemes.viridis,
-                              min_edge_size = 0.0,
-                              max_edge_size = 10.0,
-                              node_weighted = true,
-                              node_color = (:blue, 0.5),
-                              node_size = 10.0)
-
-        g = connectome.graph
-        coords = get_coords(connectome.parc)
-        positions = Point.(zip(coords[:,1], coords[:,2], coords[:,3]))
-
-        if edge_weighted
-            ew = get_edge_weight(connectome)
-            if edge_alpha
-                edge_color = Colors.alphacolor.(get(edge_map, ew), clamp.(ew, 0.2,1.0))
-            else
-                edge_color = get(edge_map, ew)
-            end
-            edge_width = clamp.(max_edge_size .* get_edge_weight(connectome), 
-                                min_edge_size, max_edge_size)
-        else
-            edge_color = [colorant"grey" for i in 1:ne(g)]
-            edge_width = fill(max_edge_size, ne(g))
-        end
-
-        if node_weighted
-            node_width = node_size .* Array(diag(degree_matrix(connectome)))
-        else    
-            node_width = fill(node_size, nv(g))
-        end
-
-        graphplot!(g,
-                   edge_width = edge_width,
-                   edge_color = edge_color,
-                   node_size = node_width,
-                   node_color = fill(node_color, nv(g));
-                   layout = _ -> positions)
-end
+function plot_cortex end
+function plot_cortex! end
+function plot_mesh end
+function plot_roi end
+function plot_roi! end
+function plot_parc end
+function plot_parc! end
+function plot_vertices end
+function plot_vertex! end
+function plot_connectome end
+function plot_connectome! end
